+++ spacekit/skopes/jwst/cal/config.py
@@ -4,7 +4,8 @@ GENKEYS = [
     "PROGRAM",  # Program number
     "OBSERVTN",  # Observation number
     "BKGDTARG",  # Background target
-    "IS_IMPRT" "VISITYPE",  # NIRSpec imprint exposure  # Visit type
+    "IS_IMPRT", # NIRSpec imprint exposure  
+    "VISITYPE", # Visit type 
     "TSOVISIT",  # Time Series Observation visit indicator
     "TARGNAME",  # Standard astronomical catalog name for target
     "TARG_RA",  # Target RA at mid time of exposure
@@ -14,8 +15,10 @@ GENKEYS = [
     "FILTER",  # Name of the filter element used
     "PUPIL",  # Name of the pupil element used
     "GRATING",  # Name of the grating element used
+    "FXD_SLIT",  # Name of fixed slit aperture used
     "EXP_TYPE",  # Type of data in the exposure
     "CHANNEL",  # Instrument channel
+    "BAND", # MRS Wavelength band
     "SUBARRAY",  # Subarray used
     "NUMDTHPT",  # Total number of points in pattern
     "GS_RA",  # guide star right ascension
@@ -59,10 +62,10 @@ COLUMN_ORDER = {
         "detector",
         "visitype",
         "filter",
+        "pupil",
         "grating",
         "subarray",
-        "bkgdtarg",
-        "is_imprt",
+        "band",
         "nexposur",
         "numdthpt",
         "targ_max_offset",
@@ -95,7 +98,7 @@ COLUMN_ORDER = {
         "subarray",
         "nexposur",
         "numdthpt",
-        "max_targ_offset",
+        "targ_max_offset",
         "offset",
         "max_offset",
         "mean_offset",
@@ -116,7 +119,7 @@ NORM_COLS = {
         "sigma1_mean",
     ],
     "SPEC": [
-        # "targ_max_offset",
+        "targ_max_offset",
         "offset",
         "max_offset",
         "mean_offset",
@@ -140,7 +143,8 @@ NORM_COLS = {
 L3_TYPES = [
     "FGS_IMAGE",
     "MIR_IMAGE",  # (TSO & Non-TSO)
-    "NRC_IMAGE" "MIR_LRS-FIXEDSLIT",
+    "NRC_IMAGE",
+    "MIR_LRS-FIXEDSLIT",
     "MIR_MRS",
     "MIR_LYOT",
     "MIR_4QPM",
@@ -335,6 +339,7 @@ KEYPAIR_DATA = {
         "NRS_MSASPEC": 39,
     },
     "channel": {"NONE": 0, "12": 1, "34": 2, "LONG": 3, "SHORT": 4},
+    "band": {"NONE": 0, "LONG": 0, "MEDIUM": 0, "SHORT": 0, 'SHORT|LONG': 0, 'SHORT|MEDIUM': 0, 'SHORT|MEDIUM|LONG': 1},
     "subarray": {
         "NONE": 0,
         "ALLSLITS": 1,